% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_bdc_p}
\alias{add_bdc_p}
\title{Shortcut for adding adjusted p-values common to biodatacore workflows.}
\usage{
add_bdc_p(data, p.value = "p.value")
}
\arguments{
\item{data}{data frame}

\item{p.value}{scalar character: name of the p-value column.}
}
\value{
data frame
}
\description{
Shortcut for adding adjusted p-values common to biodatacore workflows.
}
\seealso{
Other augmenters: \code{\link{add_nlog10_p}},
  \code{\link{add_p.adjust}}
}
