% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_nlog10_pv}
\alias{add_nlog10_pv}
\title{Adds -1*log_10() to a data frame}
\usage{
add_nlog10_pv(data, p.value = "p.value", var = NULL)
}
\arguments{
\item{data}{data frame}

\item{p.value}{scalar character: name of the p-value column.}

\item{var}{scalar character: name of the column to be created. if `NULL` then
it attempts to create its own column. To avoid accidenntally overwriting an
existing column, it is safer to always supply a name here.}
}
\value{
data frame
}
\description{
Negative log 10 pvalues are often used when comparing a large number of
pvalues. It ends up with a more natural scale, where larger values are "more
significant."
}
\seealso{
Other augmenters: \code{\link{add_p.adjust}}
}
