% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{check}
\alias{check}
\alias{check_at}
\alias{check_if}
\title{Apply a list of checks to a dataset at specified variables}
\usage{
check_at(data, .ats, .fs, individually = TRUE)

check_if(data, .ps, .fs, individually = TRUE)
}
\arguments{
\item{data}{data frame}

\item{.ats}{list. Each element of .ats can be a character vector of names or
a numeric vector of positions. Only those elements corresponding to .at
will be modified.}

\item{.fs}{list.  Each elemnt of .fs can be a single predicate function or a
formula describing such a predicate function. Must return a scalar logical.}

\item{individually}{scalar logical. If TRUE, each column in the subset is
passed to .f. if FALSE, the entire subset dataframe is passed.}

\item{.ps}{list. Each elemnt of .ps can be a single predicate function, a formula
describing such a predicate function, or a logical vector of the same
length as the number of cols of data. Only those elements where .p evaluates to TRUE will be checked}
}
\value{
list
}
\description{
check_at and check_if apply lists predicates on lists of subsets of a data
frame. They are suitable for programming and for data analysis pipelines.
}
\details{
details about form of checks
}
\note{
\code{\link[biodatacoreUtils]{zip}} is a useful function for creating
  lists of checks and subsets.
}
