% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tidiers.R
\name{contextualize.lm}
\alias{contextualize.lm}
\title{Contextualizes an lm model}
\usage{
\method{contextualize}{lm}(x, ...)
}
\arguments{
\item{...}{Arguments passed to other methods. Currently unused.}
}
\value{
data frame
}
\description{
Contextualizes an lm model
}
