% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualize.R
\name{contextualize}
\alias{contextualize}
\alias{contextualize.NULL}
\alias{contextualize.default}
\title{Construct a single row summary of a model, fit, or other object}
\usage{
contextualize(x, ...)

\method{contextualize}{NULL}(x, ...)

\method{contextualize}{default}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
data frame
}
\description{
contextualize methods always return either a one-row data frame (except on
NULL, which returns an empty data frame). It is different from
\code{broom::glance} in that it gives information about how the model was
run, instead of EG fit statistics.
}
\section{Methods (by class)}{
\itemize{
\item \code{NULL}: Returns `NULL`.

\item \code{default}: Errors and prints class of object.
}}

