% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{data_check}
\alias{data_check}
\title{Apply a list of checks to a dataset}
\usage{
data_check(data, checks, individually = TRUE)
}
\arguments{
\item{data}{data_frame}

\item{checks}{nested list of checks}
}
\value{
list
}
\description{
Apply a list of checks to a dataset
}
