% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{df_fill_na_at}
\alias{df_fill_na_at}
\title{Title}
\usage{
df_fill_na_at(data, .vars, .fill)
}
\arguments{
\item{data}{data frame}

\item{.vars}{A list of columns generated by \code{\link[dplyr]{vars}}, or a character vector of
column names, or a numeric vector of column positions.}
}
\value{
data frame
}
\description{
Title
}
