% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{df_fill_zero_at}
\alias{df_fill_zero_at}
\title{Title}
\usage{
df_fill_zero_at(data, .vars, .f)
}
\arguments{
\item{data}{data frame}

\item{.vars}{A list of columns generated by \code{\link[dplyr]{vars}}, or a character vector of
column names, or a numeric vector of column positions.}

\item{.f}{A function, formula, or scalar atomic. Must return a scalar atomic.

  If a function, it is used as is.

  If a formula, e.g. ~ .x + 2, it is converted to a function. There are three
  ways to refer to the arguments: For a single argument function, use . For a
  two argument function, use .x and .y For more arguments, use ..1, ..2, ..3
  etc.

  This syntax allows you to create very compact anonymous functions.

  If scalar character or scalar numeric, it is converted to a formula, which
  is then converted to a function as above. The end result is that the value
  is filled in in the selected locations.}
}
\value{
data frame
}
\description{
Title
}
