% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{df_log_at}
\alias{df_log_at}
\title{Title}
\usage{
df_log_at(data, .vars, ...)
}
\arguments{
\item{data}{data}

\item{.vars}{A list of columns generated by \code{\link[dplyr]{vars}}, or a character vector of
column names, or a numeric vector of column positions.}

\item{...}{arguments passed to \code{\link{log}}}
}
\value{
data frame
}
\description{
Title
}
