% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{df_scale_at}
\alias{df_scale_at}
\title{Title}
\usage{
df_scale_at(data, .p, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{arguments passed to \code{\link{scale}}}

\item{.vars}{A list of columns generated by \code{\link[dplyr]{vars}}, or a character vector of
column names, or a numeric vector of column positions..}
}
\value{
data frame
}
\description{
Title
}
