% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{df_scale_if}
\alias{df_scale_if}
\title{Title}
\usage{
df_scale_if(data, .p, ...)
}
\arguments{
\item{data}{data frame}

\item{.p}{A predicate function to be applied to the columns or a logical
vector. The variables for which .predicate is or returns TRUE are selected.
This argument is passed to rlang::as_function() and thus supports
quosure-style lambda functions and strings representing function names.}

\item{...}{arguments passed to \code{\link{scale}}}
}
\value{
data frame
}
\description{
Title
}
