% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{df_scale_numeric_vars}
\alias{df_scale_numeric_vars}
\title{Title}
\usage{
df_scale_numeric_vars(data, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{arguments passed to \code{\link{scale}}}
}
\value{
data frame
}
\description{
Title
}
