% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{f_sub}
\alias{f_sub}
\title{Substitute a list of value - replacement pairs into a list of formulas}
\usage{
f_sub(fos, subs, cross = TRUE, named = TRUE)
}
\arguments{
\item{fos}{list of formulas}

\item{subs}{named list: value - replacement pairs. The name of the list
element is the term that will be replaced. The vector of values under that
list are the terms that will be substituted in.}

\item{cross}{logical: should the substitutions be combined, applying each
value - replacement pair serially to the same formula? Or should each
substitution be applied by itself to the formula?}

\item{named}{scalar logical: Whether or not character substitutions
should be passed as names.}
}
\value{
list of formulas
}
\description{
Substitute a list of value - replacement pairs into a list of formulas
}
\seealso{
fs_sub f_subs
}
