% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{f_subs}
\alias{f_subs}
\title{Substitute a list of value - replacement pairs into a single formula}
\usage{
f_subs(fo, subs, cross = TRUE)
}
\arguments{
\item{fo}{formula}

\item{subs}{named list: value - replacement pairs. The name of the list
element is the term that will be replaced. The vector of values under that
list are the terms that will be substituted in.}

\item{cross}{logical: should the substitutions be combined, applying each
value - replacement pair serially to the same formula? Or should each
substitution be applied by itself to the formula?}
}
\value{
list of formulas
}
\description{
Substitute a list of value - replacement pairs into a single formula
}
