% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fill_zero}
\alias{fill_zero}
\alias{fill_na}
\title{Title}
\usage{
fill_zero(x, .fill)

fill_na(x, .fill)
}
\value{
atomic vector of same length as x
}
\description{
Title
}
\seealso{
Other preprocessing utilities: \code{\link{scale2}}
}
