% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lm.R
\name{guess_glm_fam}
\alias{guess_glm_fam}
\title{Attempts to guess the proper modeling family of a variable}
\usage{
guess_glm_fam(x)
}
\arguments{
\item{x}{atomic vector}
}
\value{
family object
}
\description{
Attempts to guess the proper modeling family of a variable
}
