% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{lst_to_formulas}
\alias{lst_to_formulas}
\title{Helper function to convert a list of left and right hand side modeling vars
to simple formulas}
\usage{
lst_to_formulas(lst, lhs = 1, rhs = 2)
}
\arguments{
\item{lst}{named nested list: list of length two lists.}

\item{lhs, rhs}{scalar character: the names of the lhs and rhs elements}
}
\value{
list of formulas
}
\description{
Essentially a wrapper around \code{\link{reformulate}} with checks.
}
