% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{scale2}
\alias{scale2}
\title{A wrapper around `scale` and `as.vector` suitable for use with `dplyr::mutate`}
\usage{
scale2(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a numeric matrix(like object).}

\item{center}{either a logical value or a numeric vector of length
    equal to the number of columns of \code{x}.}

\item{scale}{either a logical value or a numeric vector of length
    equal to the number of columns of \code{x}.}
}
\description{
Removes atributes that currently break `dplyr``
}
\seealso{
Other preprocessing utilities: \code{\link{fill_zero}}
}
