% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tidy2.R
\name{tidy2.lm}
\alias{tidy2.lm}
\title{Workaround and enhancers for broom::tidy.lm models}
\usage{
\method{tidy2}{lm}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, quick = FALSE, ...)
}
\arguments{
\item{x}{lm object}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{exponentiate}{whether to exponentiate the coefficient estimates
and confidence intervals (typical for logistic regression)}

\item{quick}{whether to compute a smaller and faster version, containing
only the \code{term} and \code{estimate} columns.}

\item{...}{extra arguments (not used)}
}
\description{
Handles a specific case where some coefficents are NA in the model output. It
preserves those coefficents as NA instead of removing them. Also adds
additional output. Parameters \code{conf.int:quick} are original
\code{\link[broom]{lm_tidiers}} parameters.
}
\details{
If you have missing values in your model data, you may need to refit
  the model with \code{na.action = na.exclude}. See
  \code{\link[broom]{tidy.lm}} for more information.
}
