% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2.R
\name{tidy2.NULL}
\alias{tidy2.NULL}
\title{tidy on a NULL input}
\usage{
\method{tidy2}{NULL}(x, ...)
}
\arguments{
\item{x}{A value NULL}

\item{...}{extra arguments (not used)}
}
\value{
An empty data.frame
}
\description{
tidy on a NULL input returns an empty data frame, which means it can be
combined with other data frames (treated as "empty")
}
