% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tidiers.R
\name{contextualize.glm}
\alias{contextualize.glm}
\title{Title}
\usage{
\method{contextualize}{glm}(x, ...)
}
\arguments{
\item{...}{Arguments passed to other methods. Currently unused.}
}
\value{
data frame
}
\description{
Title
}
