% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualize.R, R/lm_tidiers.R
\name{contextualize}
\alias{contextualize}
\alias{contextualize.NULL}
\alias{contextualize.default}
\alias{contextualize.lm}
\alias{contextualize.glm}
\title{Construct a single row summary of a model, fit, or other object}
\usage{
contextualize(x, ...)

\method{contextualize}{NULL}(x, ...)

\method{contextualize}{default}(x, ...)

\method{contextualize}{lm}(x, ...)

\method{contextualize}{glm}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{Arguments passed to other methods. Currently unused.}

\item{...}{Arguments passed to other methods. Currently unused.}
}
\value{
data frame

data frame

data frame
}
\description{
contextualize methods always return either a one-row data frame (except on
NULL, which returns an empty data frame). It is different from
\code{broom::glance} in that it gives information about how the model was
run, instead of EG fit statistics.

Contextualizes an lm model

Contextualizes a glm model
}
\section{Methods (by class)}{
\itemize{
\item \code{NULL}: Returns `NULL`.

\item \code{default}: Errors and prints class of object.

\item \code{lm}: 

\item \code{glm}: 
}}

